refactor(statistics): break down lib.rs into focused modules
Split the 391-line statistics service lib.rs into separate modules:
- analytics_operations.rs: user analytics and fitness aggregation logic
- daily_activity_operations.rs: daily user activity queries and parameters
- lib.rs: clean service interface with module re-exports

This improves code organization and maintainability while preserving
all existing functionality.
